chore: integration tests running with TLS when enabled
