+++ redis/tests/support/sentinel.rs
@@ -12,8 +12,12 @@ use redis::RedisResult;
 use redis::TlsMode;
 use tempfile::TempDir;
 
+use crate::support::build_single_client;
+
 use super::build_keys_and_certs_for_tls;
 use super::get_random_available_port;
+use super::is_tls_enabled;
+use super::load_certs_from_file;
 use super::Module;
 use super::RedisServer;
 use super::TlsFilePaths;
@@ -192,7 +196,10 @@ fn wait_for_replicas_to_sync(servers: &Vec<RedisServer>, masters: u16) {
 
     for cluster_index in 0..clusters {
         let master_addr = servers[cluster_index * cluster_size].connection_info();
-        let r = wait_for_master_server(|| Ok(Client::open(master_addr.clone()).unwrap()));
+        let tls_paths = &servers.get(0).unwrap().tls_paths;
+        let r = wait_for_master_server(|| {
+            Ok(build_single_client(master_addr.clone(), tls_paths).unwrap())
+        });
         if r.is_err() {
             panic!("failed waiting for master to be ready");
         }
@@ -200,7 +207,9 @@ fn wait_for_replicas_to_sync(servers: &Vec<RedisServer>, masters: u16) {
         for replica_index in 0..replicas {
             let replica_addr =
                 servers[(cluster_index * cluster_size) + 1 + replica_index].connection_info();
-            let r = wait_for_replica(|| Ok(Client::open(replica_addr.clone()).unwrap()));
+            let r = wait_for_replica(|| {
+                Ok(build_single_client(replica_addr.clone(), tls_paths).unwrap())
+            });
             if r.is_err() {
                 panic!("failed waiting for replica to be ready and in sync");
             }
@@ -358,6 +367,20 @@ impl TestSentinelContext {
     }
 
     pub fn sentinel_node_connection_info(&self) -> SentinelNodeConnectionInfo {
+        let certificates = if is_tls_enabled() {
+            let tls_files = self
+                .cluster
+                .servers
+                .get(0)
+                .unwrap()
+                .tls_paths
+                .as_ref()
+                .unwrap();
+            Some(load_certs_from_file(tls_files))
+        } else {
+            None
+        };
+
         SentinelNodeConnectionInfo {
             tls_mode: if let ConnectionAddr::TcpTls { insecure, .. } =
                 self.cluster.servers[0].client_addr()
@@ -371,7 +394,7 @@ impl TestSentinelContext {
                 None
             },
             redis_connection_info: None,
-            certificates: None,
+            certificates,
         }
     }
