+++ redis/tests/test_cluster_async.rs
@@ -18,6 +18,9 @@ use redis::{
     RedisError, RedisFuture, RedisResult, Script, Value,
 };
 
+#[cfg(feature = "tls-rustls")]
+use support::build_single_client;
+
 use crate::support::*;
 
 #[test]
@@ -250,8 +253,13 @@ async fn test_failover(env: &TestClusterContext, requests: i32, value: i32) {
             for server in env.cluster.iter_servers() {
                 let addr = server.client_addr();
 
-                let client =
-                    build_single_client(server.connection_info(), &server.tls_paths).unwrap_or_else(|e| panic!("Failed to connect to '{addr}': {e}"));
+                #[cfg(feature = "tls-rustls")]
+                let client = build_single_client(server.connection_info(), &server.tls_paths)
+                    .unwrap_or_else(|e| panic!("Failed to connect to '{addr}': {e}"));
+
+                #[cfg(not(feature = "tls-rustls"))]
+                let client = redis::Client::open(server.connection_info())
+                    .unwrap_or_else(|e| panic!("Failed to connect to '{addr}': {e}"));
 
                 let mut conn = client
                     .get_multiplexed_async_connection()
